requirements.txt
streamlit
opencv-python
mediapipe
pillow
numpy
pandas
openpyxl